Collection of updates and fixes in proxy (#48)
•	Increase version to 0.2.1.
•	Store version in created record on registration.  Version number updates in message version checks.
•	Fix Alpine image not containing Avahi/mdns
•	Fix assert when sd_stub is included.
•	Add sequence number to poll/datamessage
•	Fix webproxy --proxy command line to support curl/lws.
•	Add ability to index into text record by string (map)
•	Fix only first txt record returned issue due to use of hashset
•	Minor cleanup in lws logging

+++ src/io_types.c
@@ -416,7 +416,7 @@ int32_t io_encode_prx_socket_properties(
     __io_encode_value(ctx, int32, prx_sp, sock_type);
     __io_encode_value(ctx, int32, prx_sp, proto_type);
     __io_encode_value(ctx, uint32, prx_sp, flags);
-    __io_encode_value(ctx, uint32, prx_sp, timeout);
+    __io_encode_value(ctx, uint64, prx_sp, timeout);
     __io_encode_object(ctx, prx_socket_address, prx_sp, address);
 
     for (size_t i = 0; i < _countof(prx_sp->options); i++)
@@ -470,7 +470,7 @@ int32_t io_decode_prx_socket_properties(
     __io_decode_value(ctx, int32, prx_sp, sock_type);
     __io_decode_value(ctx, int32, prx_sp, proto_type);
     __io_decode_value(ctx, uint32, prx_sp, flags);
-    __io_decode_value(ctx, uint32, prx_sp, timeout);
+    __io_decode_value(ctx, uint64, prx_sp, timeout);
     __io_decode_object(ctx, prx_socket_address, prx_sp, address);
 
     result = io_decode_array(ctx, "options", &size, &arr);
