+++ test/ut/io_types_ut.c
@@ -1154,7 +1154,7 @@ TEST_FUNCTION(io_encode_prx_socket_properties__success_1)
     STRICT_EXPECTED_CALL_TO_ENCODE_VALUE(&k_ctx_valid, int32, &prx_sp_valid, sock_type);
     STRICT_EXPECTED_CALL_TO_ENCODE_VALUE(&k_ctx_valid, int32, &prx_sp_valid, proto_type);
     STRICT_EXPECTED_CALL_TO_ENCODE_VALUE(&k_ctx_valid, uint32, &prx_sp_valid, flags);
-    STRICT_EXPECTED_CALL_TO_ENCODE_VALUE(&k_ctx_valid, uint32, &prx_sp_valid, timeout);
+    STRICT_EXPECTED_CALL_TO_ENCODE_VALUE(&k_ctx_valid, uint64, &prx_sp_valid, timeout);
     STRICT_EXPECTED_CALL(io_encode_object(IGNORED_PTR_ARG, "address", false, IGNORED_PTR_ARG))
         .ValidateArgumentBuffer(1, &k_ctx_valid, sizeof(io_codec_ctx_t))
         .CopyOutArgumentBuffer_object(&k_ctx_valid, sizeof(io_codec_ctx_t))
@@ -1200,7 +1200,7 @@ TEST_FUNCTION(io_encode_prx_socket_properties__success_2)
     STRICT_EXPECTED_CALL_TO_ENCODE_VALUE(&k_ctx_valid, int32, &prx_sp_valid, sock_type);
     STRICT_EXPECTED_CALL_TO_ENCODE_VALUE(&k_ctx_valid, int32, &prx_sp_valid, proto_type);
     STRICT_EXPECTED_CALL_TO_ENCODE_VALUE(&k_ctx_valid, uint32, &prx_sp_valid, flags);
-    STRICT_EXPECTED_CALL_TO_ENCODE_VALUE(&k_ctx_valid, uint32, &prx_sp_valid, timeout);
+    STRICT_EXPECTED_CALL_TO_ENCODE_VALUE(&k_ctx_valid, uint64, &prx_sp_valid, timeout);
     STRICT_EXPECTED_CALL(io_encode_object(IGNORED_PTR_ARG, "address", false, IGNORED_PTR_ARG))
         .ValidateArgumentBuffer(1, &k_ctx_valid, sizeof(io_codec_ctx_t))
         .CopyOutArgumentBuffer_object(&k_ctx_valid, sizeof(io_codec_ctx_t))
@@ -1274,7 +1274,7 @@ TEST_FUNCTION(io_encode_prx_socket_properties__neg)
     STRICT_EXPECTED_CALL_TO_ENCODE_VALUE(&k_ctx_valid, int32, &prx_sp_valid, sock_type);
     STRICT_EXPECTED_CALL_TO_ENCODE_VALUE(&k_ctx_valid, int32, &prx_sp_valid, proto_type);
     STRICT_EXPECTED_CALL_TO_ENCODE_VALUE(&k_ctx_valid, uint32, &prx_sp_valid, flags);
-    STRICT_EXPECTED_CALL_TO_ENCODE_VALUE(&k_ctx_valid, uint32, &prx_sp_valid, timeout);
+    STRICT_EXPECTED_CALL_TO_ENCODE_VALUE(&k_ctx_valid, uint64, &prx_sp_valid, timeout);
     STRICT_EXPECTED_CALL(io_encode_object(IGNORED_PTR_ARG, "address", false, IGNORED_PTR_ARG))
         .ValidateArgumentBuffer(1, &k_ctx_valid, sizeof(io_codec_ctx_t)) 
         .CopyOutArgumentBuffer_object(&k_ctx_valid, sizeof(io_codec_ctx_t))
@@ -1349,7 +1349,7 @@ TEST_FUNCTION(io_decode_prx_socket_properties__success_1)
     STRICT_EXPECTED_CALL_TO_DECODE_VALUE(&k_ctx_valid, int32, &prx_sp_valid, sock_type);
     STRICT_EXPECTED_CALL_TO_DECODE_VALUE(&k_ctx_valid, int32, &prx_sp_valid, proto_type);
     STRICT_EXPECTED_CALL_TO_DECODE_VALUE(&k_ctx_valid, uint32, &prx_sp_valid, flags);
-    STRICT_EXPECTED_CALL_TO_DECODE_VALUE(&k_ctx_valid, uint32, &prx_sp_valid, timeout);
+    STRICT_EXPECTED_CALL_TO_DECODE_VALUE(&k_ctx_valid, uint64, &prx_sp_valid, timeout);
     STRICT_EXPECTED_CALL(io_decode_object(IGNORED_PTR_ARG, "address", IGNORED_PTR_ARG, IGNORED_PTR_ARG))
         .ValidateArgumentBuffer(1, &k_ctx_valid, sizeof(io_codec_ctx_t))
         .CopyOutArgumentBuffer_is_null(&k_false, sizeof(k_false))
@@ -1396,7 +1396,7 @@ TEST_FUNCTION(io_decode_prx_socket_properties__success_2)
     STRICT_EXPECTED_CALL_TO_DECODE_VALUE(&k_ctx_valid, int32, &prx_sp_valid, sock_type);
     STRICT_EXPECTED_CALL_TO_DECODE_VALUE(&k_ctx_valid, int32, &prx_sp_valid, proto_type);
     STRICT_EXPECTED_CALL_TO_DECODE_VALUE(&k_ctx_valid, uint32, &prx_sp_valid, flags);
-    STRICT_EXPECTED_CALL_TO_DECODE_VALUE(&k_ctx_valid, uint32, &prx_sp_valid, timeout);
+    STRICT_EXPECTED_CALL_TO_DECODE_VALUE(&k_ctx_valid, uint64, &prx_sp_valid, timeout);
     STRICT_EXPECTED_CALL(io_decode_object(IGNORED_PTR_ARG, "address", IGNORED_PTR_ARG, IGNORED_PTR_ARG))
         .ValidateArgumentBuffer(1, &k_ctx_valid, sizeof(io_codec_ctx_t))
         .CopyOutArgumentBuffer_is_null(&k_false, sizeof(k_false))
@@ -1473,7 +1473,7 @@ TEST_FUNCTION(io_decode_prx_socket_properties__neg)
     STRICT_EXPECTED_CALL_TO_DECODE_VALUE(&k_ctx_valid, int32, &prx_sp_valid, sock_type);
     STRICT_EXPECTED_CALL_TO_DECODE_VALUE(&k_ctx_valid, int32, &prx_sp_valid, proto_type);
     STRICT_EXPECTED_CALL_TO_DECODE_VALUE(&k_ctx_valid, uint32, &prx_sp_valid, flags);
-    STRICT_EXPECTED_CALL_TO_DECODE_VALUE(&k_ctx_valid, uint32, &prx_sp_valid, timeout);
+    STRICT_EXPECTED_CALL_TO_DECODE_VALUE(&k_ctx_valid, uint64, &prx_sp_valid, timeout);
     STRICT_EXPECTED_CALL(io_decode_object(IGNORED_PTR_ARG, "address", IGNORED_PTR_ARG, IGNORED_PTR_ARG))
         .ValidateArgumentBuffer(1, &k_ctx_valid, sizeof(io_codec_ctx_t))
         .CopyOutArgumentBuffer_is_null(&k_false, sizeof(k_false))
